Remove rand warning

+++ srcs/board/randomizer.c
@@ -36,7 +36,7 @@ static t_tile *_get_tile_to_fill(t_game *game, int rand_nb)
 static int _generate_random_valid_number(void)
 {
 	// valid number is either 2 or 4. 2 must appear more often than 4.
-	int rand_value = 5 * ((float)rand())/RAND_MAX; // Generate either 0, 1, 2, 3, or 4
+	int rand_value = 5 * ((double)rand())/RAND_MAX; // Generate either 0, 1, 2, 3, or 4
 	if (rand_value < 4) // In theorie, this give four 2 to one 4. To adjust.
 		rand_value = 2;
 	return rand_value;
@@ -53,7 +53,7 @@ void generate_random_number_in_random_empty_tile(t_game *game)
 		// mvprintw(4, 1, msg); // Message de debug temporaire, cas spécial à gérer
 		return ;
 	}
-	int rand_nb = empty_tile_nb * ((float)rand())/RAND_MAX;
+	int rand_nb = empty_tile_nb * ((double)rand())/RAND_MAX;
 	t_tile *tile_to_fill = _get_tile_to_fill(game, rand_nb);
 	if (tile_to_fill == NULL)
 	{
